Prefetch past appointments when fetching upcoming appointments
